requirements.txt
torch
torchvision
pandas
numpy
Pillow
scikit-learn
opencv-python
matplotlib
